Fix the minimum size of help text windows to the content size
This could make  help text windows bigger than the  screen, if we want
to  go there  then  using  a QLabel  sub-class  will  need to  change,
probably  by using  a  read-only QTextEdit  instead  as that  provides
scroll bars. Maybe  consider a multi-column table for  the contents as
an alternative to scroll bars.