Test version with environment variable to set Tx audio buffer size
WSJT_TX_AUDIO_BUFFER_FRAMES takes the following values:

 -1          - use Qt/system default
 0           - use 200 mS (WSJT-X default)
 +ve integer - value is number of frames at 48 kHz

-1 is  likely to  be a  good choice on  Windows and  may macOS.  0 has
proven to be good on Windows. On Linux  0 may be OK but we need to try
other values.

The value is only a hint, the  actual value used along with the period
size (the  size of each chunk  of samples requested by  the system) is
printed  in an  info level  diagnostic message  at the  start of  each
transmission.